+++ HISTORY.rst
@@ -1,14 +1,20 @@
 History
 --------
 
+0.6.x (2012-0x-xx)
+~~~~~~~~~~~~~~~~~~~
+* Support issue `#29 <https://github.com/ikeikeikeike/tastypie-queryset-client/issues/29>`_
+* Add clear method in the ManyToManyManager.
+* Add remove method in the ManyToManyManager.
+
 0.6.3 (2012-07-15)
 ~~~~~~~~~~~~~~~~~~~
 * Moved testsuite position.
 
 0.6.2 (2012-07-15)
 ~~~~~~~~~~~~~~~~~~~
-* Fixed issue #31 (Pagination Performance)
-* Support strict field in Model.
+* Fixed issue `#31 <https://github.com/ikeikeikeike/tastypie-queryset-client/issues/30>`_ (Pagination Performance)
+* `Support strict field in Model. <https://github.com/ikeikeikeike/tastypie-queryset-client/issues/28>`_
 
 0.5.2 (2012-07-10)
 ~~~~~~~~~~~~~~~~~~~